server/analyzer/requirements.txt
kafka-python==2.0.2
textblob==0.15.3
avro==1.11.3